Function matryoshka::Pull

#[no_mangle]pub unsafe extern "C" fn Pull(
    file_system: *mut FileSystem, 
    handle: *const FileHandle, 
    file_path: *const c_char
) -> *mut Status

Pull a file from the database into the virtual file system.

@param file_system A pointer to the virtual file system.

@param inner_path The inner path on the virtual file system (mind the forward slashes as separators!)

@param file_path The path on the real file system.

@return A error ocurring during operation or nullptr on success.