Crate matryoshka

Structs

FileHandle

The handle to a file.

FileSystem

Then virtual file system.

Status

The status of the operation.

Functions

Delete⚠

Delete a file. The file handle must not be used after the call but still needs to be freed.

DestroyFileHandle⚠

Destroy a file handle.

DestroyFileSystem⚠

Destroy a file system.

DestroyStatus⚠

Destroy a status.

Find⚠

Search for a specific file(s).

GetMessage⚠

Return the error message associated with a status.

GetSize⚠

Returns the size of a file.

Load⚠

Open a SQLite database containing the Matryoshka virtual file system.

Open⚠

Open a existing file on the virtual file system.

Pull⚠

Pull a file from the database into the virtual file system.

Push⚠

Push a file to the virtual file system.